refactor[core] rename permissions to core
