+++ core/engine/rego_templates/auth.rego
