+++ core/tests/input1.json
