+++ core/tests/input2.json
