+++ core/tests/input3.json
