+++ core/tests/policies/example.rego
