+++ core/tests/test-opa.sh
